+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `Add-SqlDscNode`
     - `Remove-SqlDscNode`
     - `Repair-SqlDscServer`
+    - `Complete-SqlDscImage`
 
 ### Changed